Support fixed point numbers (#6)
* add support for fixed point numbers

This entails introducing a Number type which PidController is
parametrized by, because implementing num_traits::float::Float for fixed
point types (as found in the fixed and cordic crates) seemed like too
much work at present. The crate is now parametrized by feature flags
"fixed" and "float" which determine whether to compile it for floating
point of fixed point numbers.